remote/requirements.txt
Flask
Redis
Gunicorn
python-vlc
